Error Resolved : Amount should be in double

+++ IncomeTaxDLL/IDeduction80GGA.cs
@@ -8,7 +8,7 @@ namespace IncomeTaxDLL
 {
     interface IDeduction80GGA
     {
-        int IDeduction80GGAAmount
+        double IDeduction80GGAAmount
         {
             get;
             set;
